+++ app/[month]/page.tsx
@@ -1,0 +1,7 @@
+import MonthExpenses from "@/components/pages/MonthExpenses";
+
+const page = () => {
+  return <MonthExpenses />;
+};
+
+export default page;